add convert typescript library service
