+++ tugas/service/lib/storage.ts
@@ -1,13 +1,18 @@
-const mime = require('mime-types');
-const { Client } = require('minio');
+import mime from 'mime-types';
+import { Client } from 'minio';
 
-const ERROR_REQUIRE_OBJECT_NAME = 'error wajib memasukan nama objek';
-const ERROR_FILE_NOT_FOUND = 'error file tidak ditemukan';
+export const ERROR_REQUIRE_OBJECT_NAME = 'error wajib memasukan nama objek';
+export const ERROR_FILE_NOT_FOUND = 'error file tidak ditemukan';
 
-let client;
-let bucketname;
+let client: any;
+let bucketname: string;
 
-async function connect(_bucketname, options) {
+/**
+ * Konek ke object storage
+ * @param _bucketname nama bucket
+ * @param options konfigurasi
+ */
+export async function connect(_bucketname: string, options?: any): Promise<void> {
   client = new Client({
     ...options,
     useSSL: false,
@@ -23,7 +28,11 @@ async function connect(_bucketname, options) {
   }
 }
 
-function randomFileName(mimetype) {
+/**
+ * Acak nama file
+ * @param mimetype
+ */
+export function randomFileName(mimetype: any): string {
   return (
     new Date().getTime() +
     '-' +
@@ -33,7 +42,12 @@ function randomFileName(mimetype) {
   );
 }
 
-function saveFile(file, mimetype) {
+/**
+ * Simpan file ke object storage
+ * @param file 
+ * @param mimetype 
+ */
+export function saveFile(file: any, mimetype: any): Promise<unknown> {
   const objectName = randomFileName(mimetype);
   return new Promise((resolve, reject) => {
     client.putObject(bucketname, objectName, file, (err) => {
@@ -46,7 +60,11 @@ function saveFile(file, mimetype) {
   });
 }
 
-async function readFile(objectName) {
+/**
+ * Baca file dari object storage
+ * @param objectName 
+ */
+export async function readFile(objectName: string): Promise<any> {
   if (!objectName) {
     throw ERROR_REQUIRE_OBJECT_NAME;
   }
@@ -60,11 +78,3 @@ async function readFile(objectName) {
   }
   return client.getObject(bucketname, objectName);
 }
-
-module.exports = {
-  saveFile,
-  readFile,
-  connect,
-  ERROR_REQUIRE_OBJECT_NAME,
-  ERROR_FILE_NOT_FOUND,
-};
